semantic search

+++ behaviors/behavior_magic_key.c
@@ -111,8 +111,8 @@ static bool history_read(
 static inline uint32_t encoded_from_event(
     const struct zmk_keycode_state_changed *ev) {
 
-    return ev->keycode |
-           ((uint32_t)ev->usage_page << 16);
+    return ((uint32_t)ev->keycode) |
+           (((uint32_t)ev->usage_page) << 16);
 }
 
 static bool is_modifier(
